requirements.txt
 Flask==3.0.3
Jinja2==3.1.4
itsdangerous==2.2.0
Werkzeug==3.0.3
requests==2.32.3
PyYAML==6.0.2
stripe==10.8.0
gunicorn==22.0.0
